rename train3 to train
