+++ pyterrier_dr/jpq/__init__.py
@@ -1,5 +1,5 @@
 from .index import JPQIndex
-from .train3 import JPQTrainer
+from .train import JPQTrainer
 
 from . import utils
